fix: Remove order dependency for command line options
- Use spf13/cobra to handle command line options
- Options can now be specified in any order
